+++ src/gui/gui.py
@@ -6,17 +6,18 @@ import time
 import urllib.parse
 import urllib.request
 from collections import defaultdict
-from urllib.error import URLError
 
 from PyQt6 import QtCore, QtGui, QtWidgets, uic
 import requests
 
 from peracotta import commons
-from peracotta.commons import ParserComponents, env_to_bool, make_tree
+from peracotta.commons import ParserComponents, make_tree
 from peracotta.config import conf_dir, CONFIG
 from peracotta.constants import ICON, PATH, URL, VERSION
 from peracotta.tarallo import TaralloUploadDialog, Uploader, tarallo_success_dialog
 from peracotta.peralog import logger
+
+from .widgets import JsonWidget, ErrorDialog
 from .exceptions import MissingFeaturesError
 
 from .PeraThread import PeracottaThread
@@ -43,7 +44,6 @@ class GUI(QtWidgets.QMainWindow):
     def __init__(
         self,
         app: QtWidgets.QApplication,
-        tarallo_token: str,
     ) -> None:
         super(GUI, self).__init__()
         uic.loadUi(PATH["UI"], self)
@@ -51,7 +51,6 @@ class GUI(QtWidgets.QMainWindow):
         self.uploader = None
         self.taralloDialog = None
         self.data = list(dict())
-        self.tarallo_token = tarallo_token
         self.features = dict()
         self.encountered_types_count = defaultdict(lambda: 0)
         self.active_theme = str()
@@ -189,24 +188,15 @@ class GUI(QtWidgets.QMainWindow):
 
     @staticmethod
     def backup_features_json():
-        here: str = os.path.dirname(os.path.realpath(__file__))
-        shutil.copy2(
-            os.path.join(conf_dir, "features.json"),
-            os.path.join(conf_dir, "features.json.bak"),
-        )
+        shutil.copy2(conf_dir.joinpath("features.json"), conf_dir.joinpath("features.json.bak"))
 
     @staticmethod
     def restore_features_json():
-        here = os.path.dirname(os.path.realpath(__file__))
-        shutil.move(
-            os.path.join(conf_dir, "features.json.bak"),
-            os.path.join(conf_dir, "features.json"),
-        )
+        shutil.move(conf_dir.joinpath("features.json.bak"), conf_dir.joinpath("features.json"))
 
     def load_features_file(self, auto_update: bool):
         self.features = {}
         has_file = False
-        error = None
 
         try:
             mtime = os.path.getmtime(PATH["FEATURES"])
@@ -219,7 +209,7 @@ class GUI(QtWidgets.QMainWindow):
             try:
                 response = requests.get(f"{CONFIG['TARALLO_URL']}/features.json", headers={"User-Agent": "peracotta", "Accept": "application/json"})
 
-                with open(os.path.join(conf_dir, "features.json"), "wb") as fs:
+                with open(conf_dir.joinpath("features.json"), "wb") as fs:
                     json.dump(response.json(), fs)
 
                 has_file = True
@@ -316,7 +306,7 @@ class GUI(QtWidgets.QMainWindow):
     def upload_to_tarallo(self, checkbox: bool, bulk_id=None):
         if bulk_id == "":
             bulk_id = None
-        self.uploader = Uploader(make_tree(self.data), CONFIG["TARALLO_URL"], self.tarallo_token, bulk_id, checkbox)
+        self.uploader = Uploader(make_tree(self.data), CONFIG["TARALLO_URL"], CONFIG["TARALLO_TOKEN"], bulk_id, checkbox)
         self.uploader.successEvent.connect(self.tarallo_success)
         self.uploader.failureEvent.connect(self.tarallo_failure)
         self.uploader.start()
@@ -562,31 +552,3 @@ class GUI(QtWidgets.QMainWindow):
     def closeEvent(self, a0: QtGui.QCloseEvent) -> None:
         if self.perathread.isRunning():
             self.perathread.terminate()
-
-
-class JsonWidget(QtWidgets.QDialog):
-    def __init__(self, data: list[dict], window_size: QtCore.QSize):
-        super().__init__()
-        layout = QtWidgets.QVBoxLayout()
-        text_edit = QtWidgets.QPlainTextEdit()
-        text_edit.setWordWrapMode(QtGui.QTextOption.WrapMode.NoWrap)
-        text_edit.setPlainText(f"{json.dumps(data, indent=2)}")
-        text_edit.setReadOnly(True)
-        layout.addWidget(text_edit)
-        self.setLayout(layout)
-        new_size = QtCore.QSize(int(window_size.width() * 0.8), int(window_size.height() * 0.8))
-        self.resize(new_size)
-        self.exec()
-
-
-class ErrorDialog(QtWidgets.QDialog):
-    def __init__(self, parent: QtWidgets.QMainWindow, title: str, detailed_error: str):
-        super().__init__(parent)
-        uic.loadUi(PATH["ERRORDIALOG"], self)
-        self.setWindowTitle("Error")
-        self.iconLabel = self.findChild(QtWidgets.QLabel, "iconLabel")
-        self.textLabel = self.findChild(QtWidgets.QLabel, "textLabel")
-        self.textLabel.setText(title)
-        self.errorTextEdit = self.findChild(QtWidgets.QPlainTextEdit, "errorTextEdit")
-        self.errorTextEdit.setPlainText(detailed_error)
-        self.show()
